+++ Modules/CapacityManagement/app/Services/ProjectSimulationService.php
@@ -89,7 +89,7 @@ class ProjectSimulationService
         }
 
         return (float) $project->tasks->sum(
-            fn ($task) => max(0, (int) ($task->estimated_hours ?? 0) - (int) ($task->actual_hours ?? 0))
+            fn ($task) => max(0, (float) ($task->estimated_hours ?? 0) - (float) ($task->actual_hours ?? 0))
         );
     }
